requirements.txt
fastapi[standard]
RPi.bme280